Unbreak build against Boost 1.67

+++ lib/base/timer.cpp
@@ -292,7 +292,7 @@ void Timer::TimerThreadProc()
 
 		if (wait > 0.01) {
 			/* Wait for the next timer. */
-			l_TimerCV.timed_wait(lock, boost::posix_time::milliseconds(wait * 1000));
+			l_TimerCV.timed_wait(lock, boost::posix_time::milliseconds(long(wait * 1000)));
 
 			continue;
 		}
